remove unneded error check
